feat: add send invoice stage to attach & update subscription flows
Consolidate attach/update sheets (remove V2/V3 duplicates), add shared
SendInvoiceStage component for invoice creation, and remove the
own-Stripe-account gate on the send-invoice button.

Made-with: Cursor
