+++ vite/src/components/forms/attach-v2/components/AttachFooterV3.tsx
@@ -1,8 +1,3 @@
-import {
-	Popover,
-	PopoverContent,
-	PopoverTrigger,
-} from "@/components/ui/popover";
 import { Button } from "@/components/v2/buttons/Button";
 import { SheetFooter } from "@/components/v2/sheets/SharedSheetComponents";
 import {
@@ -11,6 +6,7 @@ import {
 	TooltipTrigger,
 } from "@/components/v2/tooltips/Tooltip";
 import { useOrg } from "@/hooks/common/useOrg";
+import { useSheetStore } from "@/hooks/stores/useSheetStore";
 import { cn } from "@/lib/utils";
 import { useAttachFormContext } from "../context/AttachFormProvider";
 import { usePlanScheduleField } from "../hooks/usePlanScheduleField";
@@ -44,8 +40,9 @@ function getConfirmLabel({
 }
 
 export function AttachFooterV3() {
-	const { isPending, previewQuery, handleConfirm, handleInvoiceAttach } =
-		useAttachFormContext();
+	const { isPending, previewQuery, handleConfirm } = useAttachFormContext();
+	const { setSheet } = useSheetStore();
+	const itemId = useSheetStore((s) => s.itemId);
 
 	const { org } = useOrg();
 	const ownStripeAccount = org?.stripe_connection !== "default";
@@ -54,11 +51,15 @@ export function AttachFooterV3() {
 	const previewData = previewQuery.data;
 	const confirmLabel = getConfirmLabel({ previewData });
 
+	const isZeroAmount = previewData && previewData.total <= 0;
+
 	const invoiceDisabledReason = isEndOfCycleSelected
 		? "Invoices are not available for end of cycle changes as there is no immediate charge to invoice"
 		: !ownStripeAccount
 			? "Connect your own Stripe account to send invoices"
-			: null;
+			: isZeroAmount
+				? "Cannot send an invoice for $0 amounts. Please confirm the change instead."
+				: null;
 
 	return (
 		<SheetFooter className="flex flex-col grid-cols-1 mt-0">
@@ -71,60 +72,19 @@ export function AttachFooterV3() {
 								invoiceDisabledReason && "cursor-not-allowed",
 							)}
 						>
-							<Popover>
-								<PopoverTrigger asChild>
-									<Button
-										variant="secondary"
-										className={cn(
-											"w-full",
-											invoiceDisabledReason && "pointer-events-none opacity-50",
-										)}
-										disabled={!invoiceDisabledReason && isPending}
-									>
-										Send an Invoice
-									</Button>
-								</PopoverTrigger>
-								{!invoiceDisabledReason && (
-									<PopoverContent className="w-(--radix-popover-trigger-width) p-0">
-										<div className="flex flex-col">
-											<button
-												type="button"
-												onClick={() =>
-													handleInvoiceAttach({
-														enableProductImmediately: true,
-													})
-												}
-												className="px-4 py-3 text-left text-sm hover:bg-accent"
-											>
-												<div className="font-medium">
-													Enable plan immediately
-												</div>
-												<div className="text-xs text-muted-foreground">
-													Enable the plan immediately and redirect to Stripe to
-													finalize the invoice
-												</div>
-											</button>
-											<button
-												type="button"
-												onClick={() =>
-													handleInvoiceAttach({
-														enableProductImmediately: false,
-													})
-												}
-												className="px-4 py-3 text-left text-sm hover:bg-accent border-t"
-											>
-												<div className="font-medium">
-													Enable plan after payment
-												</div>
-												<div className="text-xs text-muted-foreground">
-													Generate an invoice link for the customer. The plan
-													will be enabled after they pay the invoice
-												</div>
-											</button>
-										</div>
-									</PopoverContent>
+							<Button
+								variant="secondary"
+								className={cn(
+									"w-full",
+									invoiceDisabledReason && "pointer-events-none opacity-50",
 								)}
-							</Popover>
+								disabled={!invoiceDisabledReason && isPending}
+								onClick={() =>
+									setSheet({ type: "attach-send-invoice", itemId })
+								}
+							>
+								Send an Invoice
+							</Button>
 						</span>
 					</TooltipTrigger>
 					{invoiceDisabledReason && (
