+++ vite/src/components/forms/attach-v2/context/AttachFormProvider.tsx
@@ -80,7 +80,13 @@ interface AttachFormContextValue {
 
 	isPending: boolean;
 	handleConfirm: () => void;
-	handleInvoiceAttach: (params: { enableProductImmediately: boolean }) => void;
+	handleInvoiceAttach: (params: {
+		enableProductImmediately: boolean;
+		finalizeInvoice: boolean;
+	}) => Promise<{
+		stripeId: string | undefined;
+		hostedInvoiceUrl: string | null | undefined;
+	}>;
 }
 
 const AttachFormReactContext = createContext<AttachFormContextValue | null>(
@@ -93,7 +99,6 @@ interface AttachFormProviderProps {
 	initialProductId?: string;
 	onPlanEditorOpen?: () => void;
 	onPlanEditorClose?: () => void;
-	onInvoiceCreated?: (invoiceId: string) => void;
 	onCheckoutRedirect?: (checkoutUrl: string) => void;
 	onSuccess?: () => void;
 	children: ReactNode;
@@ -124,7 +129,6 @@ export function AttachFormProvider({
 	initialProductId,
 	onPlanEditorOpen,
 	onPlanEditorClose,
-	onInvoiceCreated,
 	onCheckoutRedirect,
 	onSuccess,
 	children,
@@ -381,7 +385,6 @@ export function AttachFormProvider({
 	const { handleConfirm, handleInvoiceAttach, isPending } = useAttachMutation({
 		customerId,
 		buildRequestBody,
-		onInvoiceCreated,
 		onCheckoutRedirect,
 		onSuccess,
 	});
