+++ vite/src/views/customers2/components/table/customer-products/AttachProductSheetTrigger.tsx
@@ -11,7 +11,7 @@ export function AttachProductSheetTrigger() {
 	const isAttachingProduct = useIsAttachingProduct();
 
 	const handleClick = () => {
-		setSheet({ type: "attach-product-v2" });
+		setSheet({ type: "attach-product" });
 	};
 	return (
 		<Button
