+++ vite/src/views/customers2/customer/CustomerSheets.tsx
@@ -10,10 +10,8 @@ import {
 import { useIsMobile } from "@/hooks/useIsMobile";
 import { SubscriptionCancelSheet } from "@/views/customers2/components/sheets/SubscriptionCancelSheet";
 import { SubscriptionUncancelSheet } from "@/views/customers2/components/sheets/SubscriptionUncancelSheet";
-import { SubscriptionUpdateSheet2 } from "@/views/customers2/components/sheets/SubscriptionUpdateSheet2";
+import { SubscriptionUpdateSheet } from "@/views/customers2/components/sheets/SubscriptionUpdateSheet";
 import { AttachProductSheet } from "../components/sheets/AttachProductSheet";
-import { AttachProductSheetV2 } from "../components/sheets/AttachProductSheetV2";
-import { AttachProductSheetV3 } from "../components/sheets/AttachProductSheetV3";
 import { BalanceCreateSheet } from "../components/sheets/BalanceCreateSheet";
 import { BalanceDeleteSheet } from "../components/sheets/BalanceDeleteSheet";
 import { BalanceEditSheet } from "../components/sheets/BalanceEditSheet";
@@ -25,7 +23,6 @@ import { CheckBalanceSheet } from "../components/sheets/CheckBalanceSheet";
 import { InvoiceDetailSheet } from "../components/sheets/InvoiceDetailSheet";
 import { RecordUsageSheet } from "../components/sheets/RecordUsageSheet";
 import { SubscriptionDetailSheet } from "../components/sheets/SubscriptionDetailSheet";
-import { SubscriptionUpdateSheet } from "../components/sheets/SubscriptionUpdateSheet";
 import { SyncStripeSheet } from "../components/sync-stripe/SyncStripeSheet";
 import { SHEET_ANIMATION } from "./customerAnimations";
 
@@ -45,15 +42,14 @@ export function CustomerSheets() {
 	const renderSheet = () => {
 		switch (sheetType) {
 			case "attach-product":
+			case "attach-review":
+			case "attach-send-invoice":
 				return <AttachProductSheet />;
-			case "attach-product-v2":
-				return <AttachProductSheetV3 />;
 			case "subscription-detail":
 				return <SubscriptionDetailSheet />;
 			case "subscription-update":
+			case "subscription-update-send-invoice":
 				return <SubscriptionUpdateSheet />;
-			case "subscription-update-v2":
-				return <SubscriptionUpdateSheet2 />;
 			case "subscription-cancel":
 				return <SubscriptionCancelSheet />;
 			case "subscription-uncancel":
